--- components/tomba/actions/linkedin-finder/linkedin-finder.mjs
@@ -1,36 +1,0 @@
-import app from "../../tomba.app.mjs";
-
-export default {
-  key: "tomba-linkedin-finder",
-  name: "Find LinkedIn Email",
-  description:
-    "Generate or retrieve the most likely email address from a LinkedIn URL. [See the documentation](https://tomba.io/api)",
-  version: "0.0.1",
-  type: "action",
-  annotations: {
-    destructiveHint: false,
-    openWorldHint: true,
-    readOnlyHint: true,
-  },
-  props: {
-    app,
-    linkedinUrl: {
-      propDefinition: [
-        app,
-        "linkedinUrl",
-      ],
-    },
-  },
-  async run({ $ }) {
-    const response = await this.app.findLinkedIn({
-      $,
-      linkedinUrl: this.linkedinUrl,
-    });
-
-    $.export(
-      "$summary",
-      `Successfully found email from LinkedIn profile: ${this.linkedinUrl}`,
-    );
-    return response;
-  },
-};
