+++ components/tomba/actions/search-companies/search-companies.mjs
@@ -4,7 +4,7 @@ export default {
   key: "tomba-search-companies",
   name: "Search Companies",
   description:
-    "Search for companies using natural language queries or structured filters. The AI assistant will automatically generate appropriate filters from your query. [See the documentation](https://tomba.io/api)",
+    "Search for companies using natural language queries or structured filters. The AI assistant will automatically generate appropriate filters from your query. [See the documentation](https://docs.tomba.io/api/reveal#search-companies)",
   version: "0.0.1",
   type: "action",
   annotations: {
